+++ src/Agent/NewRelic/Profiler/Configuration/InstrumentationConfiguration.h
@@ -97,6 +97,11 @@ namespace NewRelic { namespace Profiler { namespace Configuration
             {
                 returnValue = true;
             }
+            else if (NewRelic::Profiler::Strings::ContainsCaseInsensitive(instrumentationXmlFilePath, _X("NewRelic.Providers.Wrapper.CastleMonoRail2.Instrumentation.xml")))
+            {
+                returnValue = true;
+            }
+
             return returnValue;
         }